refactor(kinds): home CachedFontMetrics in aether-kinds beside FontMetrics
CachedFontMetrics is pure data over FontMetrics + scale_units with no actor/FFI dependency, so it belongs in aether-kinds (usable by native and guest) rather than the aether-actor guest SDK.
