+++ crates/aether-substrate-bundle/tests/test_bench_scenario.rs
@@ -29,15 +29,14 @@
 
 use std::path::Path;
 
-use aether_actor::CachedFontMetrics;
 use aether_data::{Kind, MailboxId};
 use aether_kinds::{
-    Camera, CaptureFrame, CaptureFrameResult, CreateTexture, CreateTextureResult, Delete,
-    DeleteResult, DrawSolidQuads, DrawText, DrawTexturedQuads, DropComponent, DropResult,
-    FontMetricsRequest, FontMetricsResult, FontRef, FrameCheck, FrameCheckResult, FrameReduction,
-    FsError, List, ListResult, LoadComponent, LoadFont, LoadFontResult, LoadResult, MailEnvelope,
-    Ping, QuadScale, QuadSpace, Read, ReadResult, ReplaceComponent, ReplaceResult, SolidQuad,
-    TexturedQuad, UiBar, UiPanel, Write, WriteResult,
+    CachedFontMetrics, Camera, CaptureFrame, CaptureFrameResult, CreateTexture,
+    CreateTextureResult, Delete, DeleteResult, DrawSolidQuads, DrawText, DrawTexturedQuads,
+    DropComponent, DropResult, FontMetricsRequest, FontMetricsResult, FontRef, FrameCheck,
+    FrameCheckResult, FrameReduction, FsError, List, ListResult, LoadComponent, LoadFont,
+    LoadFontResult, LoadResult, MailEnvelope, Ping, QuadScale, QuadSpace, Read, ReadResult,
+    ReplaceComponent, ReplaceResult, SolidQuad, TexturedQuad, UiBar, UiPanel, Write, WriteResult,
 };
 use aether_math::{Mat4, Vec3};
 use aether_substrate_bundle::test_bench::{
